refactor server and protocol
