+++ src/main/scala/zio/zmx/diagnostics/nio/InetSocketAddress.scala
@@ -23,7 +23,7 @@ import zio.IO
 class InetSocketAddress private (val address: JInetSocketAddress)
 object InetSocketAddress {
 
-  def apply(host: String, port: Int): IO[Exception, InetSocketAddress] =
+  def make(host: String, port: Int): IO[Exception, InetSocketAddress] =
     IO.effect(new JInetSocketAddress(host, port))
       .refineToOrDie[Exception]
       .map(new InetSocketAddress(_))
