+++ src/main/scala/zio/zmx/diagnostics/protocol/Request.scala
@@ -1,0 +1,8 @@
+package zio.zmx.diagnostics.protocol
+
+import zio.Chunk
+
+final case class Request(
+  command: Message.Command,
+  args: Option[Chunk[String]]
+)
