+++ src/main/scala/zio/zmx/diagnostics/protocol/Response.scala
@@ -1,0 +1,8 @@
+package zio.zmx.diagnostics.protocol
+
+sealed trait Response
+
+object Response {
+  final case class Fail(error: Message.Error)  extends Response
+  final case class Success(data: Message.Data) extends Response
+}
